code/requirements.txt
flask
flask-socketio
pyserial
numpy
opencv-python       
vidgear[core]
imagezmq
paho-mqtt
flask-mqtt
redis